+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 10.0.1 - 2020/5/4
+
+- added resource info on putRender Error
+
 ## 10.0.0 - 2020/4/30
 
 - consolidate classes into core (no changes from 9.3.0)